requirement.txt
# packages in environment at C:\Users\mketk\Miniconda3\envs\math-maker:
#
# Name                    Version                   Build  Channel
ca-certificates           2019.9.11            hecc5488_0    conda-forge
certifi                   2019.9.11                py37_0    conda-forge
libiconv                  1.15              hfa6e2cd_1005    conda-forge
libxml2                   2.9.10               h9ce36c8_0    conda-forge
libxslt                   1.1.33               heafd4d3_0    conda-forge
lxml                      4.4.1            py37heafd4d3_0    conda-forge
openssl                   1.1.1d               hfa6e2cd_0    conda-forge
pip                       19.3.1                   py37_0  
python                    3.7.4                h5263a28_0  
python-docx               0.8.10                     py_0    conda-forge
setuptools                41.6.0                   py37_0  
sqlite                    3.30.1               he774522_0  
vc                        14.1                 h0510ff6_4  
vs2015_runtime            14.16.27012          hf0eaf9b_0  
wheel                     0.33.6                   py37_0  
wincertstore              0.2                      py37_0  
zlib                      1.2.11            h2fa13f4_1006    conda-forge
